Added essential emoji glue to character whitelist
u{200d} is ZWJ, the thing that combines different emojis. u{fe0f} is a
variation selector.

Only having one variation selector may become a problem, as there's 16
of them in total. As far as I can tell, only fe0f is used by emojis, so
this should work for now.

This change allows combined emojis to work, which is probably most of
them by now.